@@ -1,54 +1,48 @@
 [package]
 name = "bestool-psql"
-version = "0.2.1"
-edition = "2021"
+version = "1.0.0"
+edition = "2024"
 authors = ["Félix Saparelli <felix@passcod.name>", "BES Developers <contact@bes.au>"]
 license = "GPL-3.0-or-later"
-description = "Custom psql wrapper for BES (bestool dependency)"
-repository = "https://github.com/beyondessential/bestool"
+description = "psql-inspired client for PostgreSQL"
+repository = "https://github.com/beyondessential/bestool/tree/main/crates/psql"
 
 [[bin]]
 name = "bestool-psql"
 path = "src/main.rs"
 required-features = ["cli"]
 
-[[bin]]
-name = "psql-history"
-path = "src/bin/psql-history.rs"
-required-features = ["cli"]
-
 [dependencies]
 clap = { version = "4.5.47", features = ["cargo", "derive", "wrap_help"], optional = true }
+comfy-table = "7.2.1"
+crossterm = "0.29.0"
+ctrlc = "3.4.4"
+dirs = "6.0.0"
 edit = "0.1.5"
+hex = "0.4.3"
 jiff = "0.2.15"
 lloggs = { version = "1.1.0", optional = true }
 miette = "7.6.0"
-portable-pty = "0.9.0"
-rand = "0.9.2"
+mobc = "0.9.0"
+mobc-postgres = { version = "0.9.0", features = ["with-serde_json-1"] }
 redb = "3.1.0"
-regex = "1.12.2"
-rustyline = "17.0.2"
+rustls = { version = "0.23", default-features = false, features = ["ring", "logging", "std", "tls12"] }
+rustls-native-certs = "0.8.2"
+rustyline = { version = "17.0.2", features = ["signal-hook"] }
 serde = { version = "1.0.228", features = ["derive"] }
 serde_json = "1.0.145"
-strip-ansi-escapes = "0.2.1"
+supports-unicode = "3.0.0"
 syntect = { version = "5.3.0", default-features = false, features = ["parsing", "regex-fancy", "default-themes", "default-syntaxes"] }
 tempfile = "3.21.0"
-terminal_size = "0.4.0"
-thiserror = "2.0.16"
 tokio = { version = "1.47.1", features = ["full"] }
+tokio-postgres = { version = "0.7.15", features = ["with-jiff-0_2", "with-serde_json-1"] }
+tokio-postgres-rustls = "0.13.0"
 tracing = "0.1.41"
+winnow = "0.7.13"
 
 [target.'cfg(unix)'.dependencies]
 libc = "0.2.177"
 signal-hook = "0.3.18"
-
-[target.'cfg(windows)'.dependencies]
-windows-sys = { version = "0.61.2", features = [
-    "Win32_System_Console",
-    "Win32_Foundation",
-    "Win32_Storage_FileSystem",
-] }
-
 
 [features]
 default = ["cli"]
