@@ -1,6 +1,8 @@
 use clap::{Parser, Subcommand};
 use miette::Result;
 
+use crate::args::Args;
+
 use super::Context;
 
 /// File utilities.
@@ -12,8 +14,8 @@ pub struct FileArgs {
 }
 
 super::subcommands! {
-	[Context<FileArgs> => {|ctx: Context<FileArgs>| -> Result<(Action, Context<FileArgs>)> {
-		Ok((ctx.args_top.action.clone(), ctx.with_sub(())))
+	[Context<Args, FileArgs> => {|ctx: Context<Args, FileArgs>| -> Result<(Action, Context<FileArgs>)> {
+		Ok((ctx.args_sub.action.clone(), ctx.push(())))
 	}}](with_sub)
 
 	join => Join(JoinArgs),
