@@ -1,6 +1,8 @@
 use clap::{Parser, Subcommand};
 use miette::Result;
 
+use crate::args::Args;
+
 use super::Context;
 
 /// Tamanu Iti subcommands.
@@ -12,8 +14,8 @@ pub struct ItiArgs {
 }
 
 super::subcommands! {
-	[Context<ItiArgs> => {|ctx: Context<ItiArgs>| -> Result<(Action, Context<ItiArgs>)> {
-		Ok((ctx.args_top.action.clone(), ctx.with_sub(())))
+	[Context<Args, ItiArgs> => {|ctx: Context<Args, ItiArgs>| -> Result<(Action, Context<ItiArgs>)> {
+		Ok((ctx.args_sub.action.clone(), ctx.push(())))
 	}}](with_sub)
 
 	#[cfg(feature = "iti-battery")]
