@@ -1,27 +1,15 @@
-//! OTS (Over The Shoulder) prompting with history support
-
-use crate::history::History;
 use miette::{IntoDiagnostic, Result};
-use redb::{Database, ReadableDatabase, ReadableTable};
-use rustyline::history::{History as HistoryTrait, MemHistory};
-use rustyline::{Config, Editor};
-use std::path::Path;
-use std::sync::{Arc, RwLock};
-use tracing::{debug, warn};
+use rustyline::{
+	history::{History as HistoryTrait, MemHistory},
+	Config, Editor,
+};
+use tracing::debug;
+
+use crate::audit::Audit;
 
 /// Prompt for OTS value with rustyline and history from previous OTS values
-///
-/// This can either open the history database from a path, or use a shared database handle
-pub fn prompt_for_ots(history_path: &Path) -> Result<String> {
-	prompt_for_ots_with_db(None, Some(history_path))
-}
-
-/// Prompt for OTS value with rustyline and history from a shared database handle
-pub fn prompt_for_ots_with_db(
-	db: Option<Arc<RwLock<Database>>>,
-	history_path: Option<&Path>,
-) -> Result<String> {
-	let ots_history = load_ots_history(db, history_path)?;
+pub fn prompt_for_ots(audit: &Audit) -> Result<String> {
+	let ots_history = load_ots_history(audit)?;
 
 	let mut rl: Editor<(), MemHistory> = Editor::with_history(
 		Config::builder()
@@ -60,35 +48,10 @@ pub fn prompt_for_ots_with_db(
 	}
 }
 
-/// Load unique OTS values from the history database
-fn load_ots_history(
-	db: Option<Arc<RwLock<Database>>>,
-	history_path: Option<&Path>,
-) -> Result<Vec<String>> {
-	let entries = if let Some(db) = db {
-		load_entries_from_db(&db)?
-	} else if let Some(path) = history_path {
-		let history = match History::open(path) {
-			Ok(h) => h,
-			Err(e) => {
-				warn!("could not open history database for OTS values: {}", e);
-				return Ok(Vec::new());
-			}
-		};
+/// Load unique OTS values from the audit database
+fn load_ots_history(audit: &Audit) -> Result<Vec<String>> {
+	let entries = audit.list()?;
 
-		match history.list() {
-			Ok(e) => e,
-			Err(e) => {
-				warn!("could not read history entries for OTS values: {}", e);
-				return Ok(Vec::new());
-			}
-		}
-	} else {
-		warn!("no database handle or path provided");
-		return Ok(Vec::new());
-	};
-
-	// Collect unique OTS values (most recent order preserved by using Vec and dedup)
 	let mut ots_values = Vec::new();
 	let mut seen = std::collections::HashSet::new();
 	for (_timestamp, entry) in entries.into_iter().rev() {
@@ -101,24 +64,4 @@ fn load_ots_history(
 
 	debug!(count = ots_values.len(), "loaded OTS history");
 	Ok(ots_values)
-}
-
-/// Load entries directly from a database handle
-fn load_entries_from_db(
-	db: &Arc<RwLock<Database>>,
-) -> Result<Vec<(u64, crate::history::HistoryEntry)>> {
-	use crate::history::{HistoryEntry, HISTORY_TABLE};
-	use miette::IntoDiagnostic;
-
-	let read_txn = db.read().unwrap().begin_read().into_diagnostic()?;
-	let table = read_txn.open_table(HISTORY_TABLE).into_diagnostic()?;
-
-	let mut entries = Vec::new();
-	for item in table.iter().into_diagnostic()? {
-		let (timestamp, json) = item.into_diagnostic()?;
-		let entry: HistoryEntry = serde_json::from_str(json.value()).into_diagnostic()?;
-		entries.push((timestamp.value(), entry));
-	}
-
-	Ok(entries)
 }
