@@ -10,9 +10,9 @@ pub(crate) fn reset_sigint() {
 	SIGINT_RECEIVED.store(false, Ordering::SeqCst);
 }
 
-pub fn register_sigint_handler() -> Result<(), Box<dyn std::error::Error>> {
+pub fn register_sigint_handler() -> miette::Result<()> {
 	ctrlc::set_handler(move || {
 		SIGINT_RECEIVED.store(true, Ordering::SeqCst);
-	})?;
-	Ok(())
+	})
+	.map_err(|e| miette::miette!("Failed to register Ctrl-C handler: {e}"))
 }
